examples/riva_speech_langchain/pyproject.toml
[project]
name = "riva-speech-langchain-example"
version = "0.1.0"
description = "NVIDIA ACE Pipecat Speech only Examples"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "langchain>=0.3.19",
 "langchain-community>=0.3.18",
 "langchain-openai>=0.3.6",
 "nvidia-pipecat",
]

[tool.uv.sources]
torch = { index = "pytorch" }
nvidia-pipecat = { path = "../../.", editable = true }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
